Improved parser, added branching support
this allows a few new tricks:
- showing only usages of the most likely candidates on invalid syntax (e.g. typing 'time set' only shows the usage for 'time set' and not 'time add' etc)
- non-brute-force resolving of subcommands when the first parameter in a branched executor is a literal (common case)
- avoiding repeated parsing of parameters shared by multiple overloads, if branching is set up properly

There are still some outstanding issues, such as giving too many arguments to a command currently shows the longest overload that would've worked,
but if too many arguments were given, it's anyone's guess what the user's intention was. Not clear what to do in this case.

It's also possible we might not want to dump the entire usage message(s) into the user's chat window in some cases.
Sometimes plugins create very complex command trees, and it would suck to get spammed with dozens of usage messages for the sake of one mistake.
It might be nice to have usage support for branches to allow only showing the next argument, or something of that nature.